fix: remove the flight reference from the booking db
This was a bug that was left untouched from the moment that i went from
a one to one relation ship, to a has and belong to many that was more
appropriate to reflect the relationship that was used in the logic.

when sketching out the blueprint for the relations between booking
flight, i initially created a one to one relation just for boilerplate,
and then when i actually when about building the relationships that i
intended i introduced a bug where the migration was still expecting to
point to the flight reference even tough that doesn't make sense in a has
and belong to many.